+++ tests/test_repo_manage.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 
 import os
+import time
 import logging
 
 class MyLogger(logging.getLoggerClass()):